Added GUI
Added a GUI using the Laterna library. Laterna is a ncurses-like
library that creates a terminal window, if necessary, to display
the UI.
The UI also calculates the time it takes to calculate a new population
from the old and displays those values. Not yet capable of accepting
user input, but that is forthcoming.